feat: Set up for swiss pairing rounds
This is a massive commit unfortunately. It adds a ReportMatch function
that is accessible on the view where a winner from a match up in a
Swiss Pairing can be chosen by clicking on the name.

The other major and not fully implemented addition is to have a temp
global array hold match objects for each match that is supposed to be
played in the current round. This prevents new swiss pairings to be
made until all the current games have been reported. This is just
the groundwork for this feature though.

Current state: If there are no matches to be played, assume we need to
load a new set of Swiss Pairings. Using the SwissPairings from the
tournament.py file, create a temporary object that holds the current
matches to be played with some supporting attributes. Also, specify that
each of these pairings corresponds to a match to be played Then create
the HTML list of li elements as fresh matches to be played.

If there are matches to be played. Then load each match and
appropriately show completed matches with the winner present and
incomplete matches with a form to choose a winner.