fix: Clean up matches when python stores nothing
If there are no matches stored in the python code, the matches stored
in the database are leftover from a previously played tournament that
was not stored and can be removed.

WARNING: This is a naive and lazy implementation done for this test
project and should NOT be used in a real production application since
it can lead to frustration over accidental closes / crashes causing
loss of information. Ideally, the matches should just be stored in
the database so that they are less ephemeral and can be carried over
to finish tournaments between rounds.

One way to organizing this is to have a tournament_match table that
stores matches related to a tournament in the approparite order. This
would require also storing whether a tournament is over... how many
pairs are in each tournament... what players are in the tournament.
The complexity of this in light of this being just a basic showcase
project for basic SQL / backend python is what made me choose this
non-solution. Also, it must be admitted that I haven't thought about
this and a simpler solution exists that would require less work than
I had just described.